Move completerr functions and const classes
