+++ kit/hekit.py
@@ -14,9 +14,8 @@ from pathlib import Path
 from typing import Tuple
 
 from kit.utils.subparsers import register_subparser, validate_input
-from kit.utils.constants import Constants
-from kit.utils.tab_completion import enable_tab_completion
-from kit.commands.plugin import get_plugins_by_state
+from kit.utils.constants import Constants, PluginsConfig
+from kit.utils.tab_completion import enable_tab_completion, get_plugins_by_state
 
 if sys.version_info < (3, 8):
     print("Intel HE Toolkit requires Python version 3.8 or above", file=sys.stderr)
@@ -45,7 +44,7 @@ def parse_cmdline() -> Tuple:
     # create subparsers for each command
     subparsers = parser.add_subparsers(help="sub-command help")
     register_subparser(subparsers, ["commands", "tools"], hekit_root_dir / "kit")
-    register_subparser(subparsers, get_plugins_by_state(), Constants.plugins_root_dir)
+    register_subparser(subparsers, get_plugins_by_state(), PluginsConfig.ROOT_DIR)
 
     # try to enable tab completion
     enable_tab_completion(parser)
